Revert "Add long and varchar enum types"
This reverts commit 7ced455b02032a99b7ce8991412897baba60d14d.
