+++ presto-common/src/main/java/com/facebook/presto/common/type/ParameterKind.java
@@ -23,9 +23,7 @@ public enum ParameterKind
     TYPE(Optional.of("TYPE_SIGNATURE")),
     NAMED_TYPE(Optional.of("NAMED_TYPE_SIGNATURE")),
     LONG(Optional.of("LONG_LITERAL")),
-    VARIABLE(Optional.empty()),
-    LONG_ENUM(Optional.of("LONG_ENUM")),
-    VARCHAR_ENUM(Optional.of("VARCHAR_ENUM"));
+    VARIABLE(Optional.empty());
 
     // TODO: drop special serialization code as soon as all clients
     //       migrate to version which can deserialize new format.
