+++ presto-common/src/main/java/com/facebook/presto/common/type/TypeSignature.java
@@ -13,15 +13,13 @@
  */
 package com.facebook.presto.common.type;
 
-import com.facebook.presto.common.type.LongEnumType.LongEnumMap;
-import com.facebook.presto.common.type.VarcharEnumType.VarcharEnumMap;
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonValue;
 
 import java.util.ArrayList;
-import java.util.HashMap;
 import java.util.HashSet;
 import java.util.List;
+import java.util.Locale;
 import java.util.Map;
 import java.util.Objects;
 import java.util.Optional;
@@ -29,13 +27,11 @@ import java.util.Set;
 import java.util.TreeMap;
 import java.util.TreeSet;
 import java.util.regex.Pattern;
-import java.util.stream.Collectors;
 
 import static java.lang.Character.isDigit;
 import static java.lang.String.format;
 import static java.util.Arrays.asList;
 import static java.util.Collections.unmodifiableList;
-import static java.util.Locale.ENGLISH;
 
 public class TypeSignature
 {
@@ -48,9 +44,6 @@ public class TypeSignature
             new TreeMap<>(String.CASE_INSENSITIVE_ORDER);
     private static final Set<String> SIMPLE_TYPE_WITH_SPACES =
             new TreeSet<>(String.CASE_INSENSITIVE_ORDER);
-
-    private static final String LONG_ENUM_PREFIX = "enum:bigint";
-    private static final String VARCHAR_ENUM_PREFIX = "enum:varchar";
 
     static {
         BASE_NAME_ALIAS_TO_CANONICAL.put("int", StandardTypes.INTEGER);
@@ -107,21 +100,6 @@ public class TypeSignature
         return calculated;
     }
 
-    public boolean isEnum()
-    {
-        return isLongEnum() || isVarcharEnum();
-    }
-
-    public boolean isLongEnum()
-    {
-        return parameters.size() == 1 && parameters.get(0).isLongEnum();
-    }
-
-    public boolean isVarcharEnum()
-    {
-        return parameters.size() == 1 && parameters.get(0).isVarcharEnum();
-    }
-
     @JsonCreator
     public static TypeSignature parseTypeSignature(String signature)
     {
@@ -137,7 +115,7 @@ public class TypeSignature
             checkArgument(!literalCalculationParameters.contains(signature), "Bad type signature: '%s'", signature);
             return new TypeSignature(canonicalizeBaseName(signature), new ArrayList<>());
         }
-        if (signature.toLowerCase(ENGLISH).startsWith(StandardTypes.ROW + "(")) {
+        if (signature.toLowerCase(Locale.ENGLISH).startsWith(StandardTypes.ROW + "(")) {
             return parseRowTypeSignature(signature, literalCalculationParameters);
         }
 
@@ -145,12 +123,8 @@ public class TypeSignature
         List<TypeSignatureParameter> parameters = new ArrayList<>();
         int parameterStart = -1;
         int bracketCount = 0;
-        int parameterEnd = -1;
 
         for (int i = 0; i < signature.length(); i++) {
-            if (i < parameterEnd) {
-                continue;
-            }
             char c = signature.charAt(i);
             // TODO: remove angle brackets support once ROW<TYPE>(name) will be dropped
             // Angle brackets here are checked not for the support of ARRAY<> and MAP<>
@@ -177,9 +151,6 @@ public class TypeSignature
                     }
                 }
             }
-            else if (isEnumMapStart(signature, i)) {
-                parameterEnd = parseEnumMap(signature, i).mapEndIndex;
-            }
             else if (c == ',') {
                 if (bracketCount == 1) {
                     checkArgument(parameterStart >= 0, "Bad type signature: '%s'", signature);
@@ -190,133 +161,6 @@ public class TypeSignature
         }
 
         throw new IllegalArgumentException(format("Bad type signature: '%s'", signature));
-    }
-
-    private static class EnumMapParsingData
-    {
-        final int mapEndIndex;
-        private final Map<String, String> map;
-        final boolean isLongEnum;
-
-        EnumMapParsingData(int mapEndIndex, Map<String, String> map, boolean isLongEnum)
-        {
-            this.mapEndIndex = mapEndIndex;
-            this.map = map;
-            this.isLongEnum = isLongEnum;
-        }
-
-        LongEnumMap getLongEnumMap()
-        {
-            checkArgument(isLongEnum, "Invalid enum map format");
-            return new LongEnumMap(
-                    map.entrySet().stream()
-                            .collect(Collectors.toMap(Map.Entry::getKey, e -> Long.parseLong(e.getValue()))));
-        }
-
-        VarcharEnumMap getVarcharEnumMap()
-        {
-            checkArgument(!isLongEnum, "Invalid enum map format");
-            return new VarcharEnumMap(map);
-        }
-    }
-
-    private static boolean isEnumMapStart(String signature, int startIndex)
-    {
-        String suffix = signature.substring(startIndex).toLowerCase(ENGLISH);
-        return suffix.startsWith(LONG_ENUM_PREFIX + "{") || suffix.startsWith(VARCHAR_ENUM_PREFIX + "{");
-    }
-
-    private enum EnumMapParsingState
-    {
-        EXPECT_KEY,
-        IN_KEY,
-        IN_KEY_ESCAPE,
-        EXPECT_COLON,
-        EXPECT_VALUE,
-        IN_NUM_VALUE,
-        IN_STR_VALUE,
-        IN_STR_VALUE_ESCAPE,
-        EXPECT_COMMA_OR_CLOSING_BRACKET
-    }
-
-    private static EnumMapParsingData parseEnumMap(String signature, int startIndex)
-    {
-        EnumMapParsingState state = EnumMapParsingState.EXPECT_KEY;
-        boolean isLongEnum = signature.substring(startIndex).trim().toLowerCase(ENGLISH).startsWith(LONG_ENUM_PREFIX);
-        int openBracketIndex = startIndex + (isLongEnum ? LONG_ENUM_PREFIX.length() : VARCHAR_ENUM_PREFIX.length()) + 1;
-        String key = null;
-        StringBuilder keyOrValue = new StringBuilder();
-        Map<String, String> map = new HashMap<>();
-
-        for (int i = openBracketIndex; i < signature.length(); i++) {
-            char c = signature.charAt(i);
-            if (state == EnumMapParsingState.IN_KEY_ESCAPE) {
-                state = EnumMapParsingState.IN_KEY;
-                keyOrValue.append(c);
-            }
-            if (state == EnumMapParsingState.IN_STR_VALUE_ESCAPE) {
-                state = EnumMapParsingState.IN_STR_VALUE;
-                keyOrValue.append(c);
-            }
-            else if (c == '"') {
-                if (state == EnumMapParsingState.EXPECT_KEY) {
-                    state = EnumMapParsingState.IN_KEY;
-                }
-                else if (state == EnumMapParsingState.EXPECT_VALUE) {
-                    if (isLongEnum) {
-                        throw new IllegalStateException("Unexpected varchar value in numeric enum signature");
-                    }
-                    state = EnumMapParsingState.IN_STR_VALUE;
-                }
-                else if ((state == EnumMapParsingState.IN_KEY || state == EnumMapParsingState.IN_STR_VALUE)
-                        && i + 1 < signature.length() && signature.charAt(i + 1) == '"') {
-                    state = state == EnumMapParsingState.IN_KEY ? EnumMapParsingState.IN_KEY_ESCAPE : EnumMapParsingState.IN_STR_VALUE_ESCAPE;
-                }
-                else if (state == EnumMapParsingState.IN_KEY) {
-                    state = EnumMapParsingState.EXPECT_COLON;
-                }
-                else if (state == EnumMapParsingState.IN_STR_VALUE) {
-                    state = EnumMapParsingState.EXPECT_COMMA_OR_CLOSING_BRACKET;
-                }
-                else {
-                    throw new IllegalStateException("Cannot parse enum signature");
-                }
-            }
-            else if (state == EnumMapParsingState.IN_KEY || state == EnumMapParsingState.IN_STR_VALUE) {
-                keyOrValue.append(c);
-            }
-            else if (c == ':' && state == EnumMapParsingState.EXPECT_COLON) {
-                key = keyOrValue.toString();
-                keyOrValue = new StringBuilder();
-                state = EnumMapParsingState.EXPECT_VALUE;
-            }
-            else if ((Character.isDigit(c) || c == '-') && state == EnumMapParsingState.EXPECT_VALUE) {
-                if (!isLongEnum) {
-                    throw new IllegalStateException("Unexpected numeric value in varchar enum signature");
-                }
-                state = EnumMapParsingState.IN_NUM_VALUE;
-                keyOrValue.append(c);
-            }
-            else if (Character.isDigit(c) && state == EnumMapParsingState.IN_NUM_VALUE) {
-                keyOrValue.append(c);
-            }
-            else if ((c == ',' || c == '}') && (state == EnumMapParsingState.EXPECT_COMMA_OR_CLOSING_BRACKET || state == EnumMapParsingState.IN_NUM_VALUE)) {
-                if (key == null) {
-                    throw new IllegalStateException("Cannot parse enum signature");
-                }
-                map.put(key, keyOrValue.toString());
-                if (c == '}') {
-                    return new EnumMapParsingData(i, map, isLongEnum);
-                }
-                key = null;
-                keyOrValue = new StringBuilder();
-                state = EnumMapParsingState.EXPECT_KEY;
-            }
-            else if (!Character.isWhitespace(c)) {
-                throw new IllegalStateException("Cannot parse enum signature");
-            }
-        }
-        throw new IllegalStateException("Cannot parse enum signature");
     }
 
     private enum RowTypeSignatureParsingState
@@ -331,7 +175,7 @@ public class TypeSignature
 
     private static TypeSignature parseRowTypeSignature(String signature, Set<String> literalParameters)
     {
-        checkArgument(signature.toLowerCase(ENGLISH).startsWith(StandardTypes.ROW + "("), "Not a row type signature: '%s'", signature);
+        checkArgument(signature.toLowerCase(Locale.ENGLISH).startsWith(StandardTypes.ROW + "("), "Not a row type signature: '%s'", signature);
 
         RowTypeSignatureParsingState state = RowTypeSignatureParsingState.START_OF_FIELD;
         int bracketLevel = 1;
@@ -474,18 +318,6 @@ public class TypeSignature
         else if (literalCalculationParameters.contains(parameterName)) {
             return TypeSignatureParameter.of(parameterName);
         }
-        else if (isEnumMapStart(signature, begin)) {
-            if (!parameterName.endsWith("}")) {
-                throw new IllegalStateException("Cannot parse enum signature");
-            }
-            EnumMapParsingData enumMapData = parseEnumMap(signature, begin);
-            if (enumMapData.isLongEnum) {
-                return TypeSignatureParameter.of(enumMapData.getLongEnumMap());
-            }
-            else {
-                return TypeSignatureParameter.of(enumMapData.getVarcharEnumMap());
-            }
-        }
         else {
             return TypeSignatureParameter.of(parseTypeSignature(parameterName, literalCalculationParameters));
         }
@@ -562,13 +394,13 @@ public class TypeSignature
 
         TypeSignature other = (TypeSignature) o;
 
-        return Objects.equals(this.base.toLowerCase(ENGLISH), other.base.toLowerCase(ENGLISH)) &&
+        return Objects.equals(this.base.toLowerCase(Locale.ENGLISH), other.base.toLowerCase(Locale.ENGLISH)) &&
                 Objects.equals(this.parameters, other.parameters);
     }
 
     @Override
     public int hashCode()
     {
-        return Objects.hash(base.toLowerCase(ENGLISH), parameters);
+        return Objects.hash(base.toLowerCase(Locale.ENGLISH), parameters);
     }
 }
