+++ presto-common/src/main/java/com/facebook/presto/common/type/TypeSignatureParameter.java
@@ -13,9 +13,6 @@
  */
 package com.facebook.presto.common.type;
 
-import com.facebook.presto.common.type.LongEnumType.LongEnumMap;
-import com.facebook.presto.common.type.VarcharEnumType.VarcharEnumMap;
-
 import java.util.Objects;
 import java.util.Optional;
 
@@ -45,16 +42,6 @@ public class TypeSignatureParameter
     public static TypeSignatureParameter of(String variable)
     {
         return new TypeSignatureParameter(ParameterKind.VARIABLE, variable);
-    }
-
-    public static TypeSignatureParameter of(LongEnumMap enumMap)
-    {
-        return new TypeSignatureParameter(ParameterKind.LONG_ENUM, enumMap);
-    }
-
-    public static TypeSignatureParameter of(VarcharEnumMap enumMap)
-    {
-        return new TypeSignatureParameter(ParameterKind.VARCHAR_ENUM, enumMap);
     }
 
     private TypeSignatureParameter(ParameterKind kind, Object value)
@@ -94,16 +81,6 @@ public class TypeSignatureParameter
         return kind == ParameterKind.VARIABLE;
     }
 
-    public boolean isLongEnum()
-    {
-        return kind == ParameterKind.LONG_ENUM;
-    }
-
-    public boolean isVarcharEnum()
-    {
-        return kind == ParameterKind.VARCHAR_ENUM;
-    }
-
     private <A> A getValue(ParameterKind expectedParameterKind, Class<A> target)
     {
         if (kind != expectedParameterKind) {
@@ -132,16 +109,6 @@ public class TypeSignatureParameter
         return getValue(ParameterKind.VARIABLE, String.class);
     }
 
-    public LongEnumMap getLongEnumMap()
-    {
-        return getValue(ParameterKind.LONG_ENUM, LongEnumMap.class);
-    }
-
-    public VarcharEnumMap getVarcharEnumMap()
-    {
-        return getValue(ParameterKind.VARCHAR_ENUM, VarcharEnumMap.class);
-    }
-
     public Optional<TypeSignature> getTypeSignatureOrNamedTypeSignature()
     {
         switch (kind) {
@@ -162,8 +129,6 @@ public class TypeSignatureParameter
             case NAMED_TYPE:
                 return getNamedTypeSignature().getTypeSignature().isCalculated();
             case LONG:
-            case LONG_ENUM:
-            case VARCHAR_ENUM:
                 return false;
             case VARIABLE:
                 return true;
