+++ presto-common/src/main/java/com/facebook/presto/common/type/TypeUtils.java
@@ -19,14 +19,6 @@ import com.facebook.presto.common.block.BlockBuilder;
 import io.airlift.slice.Slice;
 import io.airlift.slice.Slices;
 
-import java.util.Map;
-import java.util.Objects;
-import java.util.stream.Collectors;
-
-import static java.util.Locale.ENGLISH;
-import static java.util.stream.Collectors.toMap;
-import static java.util.stream.Collectors.toSet;
-
 public final class TypeUtils
 {
     public static final int NULL_HASH_CODE = 0;
@@ -108,27 +100,5 @@ public final class TypeUtils
         if (isNull) {
             throw new NotSupportedException(errorMsg);
         }
-    }
-
-    static void validateEnumMap(Map<String, ?> enumMap)
-    {
-        if (enumMap.containsKey(null)) {
-            throw new IllegalArgumentException("Enum cannot contain null key");
-        }
-        int nUniqueAndNotNullValues = enumMap.values().stream()
-                .filter(Objects::nonNull).collect(toSet()).size();
-        if (nUniqueAndNotNullValues != enumMap.size()) {
-            throw new IllegalArgumentException("Enum cannot contain null or duplicate values");
-        }
-        int nCaseInsensitiveKeys = enumMap.keySet().stream().map(k -> k.toUpperCase(ENGLISH)).collect(Collectors.toSet()).size();
-        if (nCaseInsensitiveKeys != enumMap.size()) {
-            throw new IllegalArgumentException("Enum cannot contain case-insensitive duplicate keys");
-        }
-    }
-
-    static <V> Map<String, V> normalizeEnumMap(Map<String, V> entries)
-    {
-        return entries.entrySet().stream()
-                .collect(toMap(e -> e.getKey().toUpperCase(ENGLISH), Map.Entry::getValue));
     }
 }
