+++ presto-common/src/test/java/com/facebook/presto/common/type/TestTypeSignature.java
@@ -13,10 +13,7 @@
  */
 package com.facebook.presto.common.type;
 
-import com.facebook.presto.common.type.LongEnumType.LongEnumMap;
-import com.facebook.presto.common.type.VarcharEnumType.VarcharEnumMap;
 import com.google.common.collect.ImmutableList;
-import com.google.common.collect.ImmutableMap;
 import com.google.common.collect.ImmutableSet;
 import org.testng.annotations.Test;
 
@@ -292,53 +289,6 @@ public class TestTypeSignature
         assertFalse(parseTypeSignature("row(a decimal(2,1),b decimal(3,2))").isCalculated());
     }
 
-    @Test
-    public void testEnumSignature()
-    {
-        assertEquals(
-                parseTypeSignature("test_enum(enum:varchar{\"test\" :\"\"\"\", \"hello\": \" \" , \"a\":\"}{{\" })"),
-                new VarcharEnumType("test_enum", new VarcharEnumMap(ImmutableMap.of("a", "}{{", "hello", " ", "test", "\""))).getTypeSignature());
-
-        assertEquals(
-                parseTypeSignature("test_enum(enum:varchar{\"my  key\" :\"मूल्य\"})"),
-                new VarcharEnumType("test_enum", new VarcharEnumMap(ImmutableMap.of("my  key", "मूल्य"))).getTypeSignature());
-
-        assertEquals(
-                parseTypeSignature("other_enum(ENUM:bigint{\"hello\" :  -5, \"AaA\"  : 9999 })"),
-                new LongEnumType("other_enum", new LongEnumMap(ImmutableMap.of("hello", -5L, "AAA", 9999L))).getTypeSignature());
-
-        assertEquals(
-                parseTypeSignature("my_enum(enum:varchar{\"))(\" :\"){}\"})"),
-                new VarcharEnumType("my_enum", new VarcharEnumMap(ImmutableMap.of("))(", "){}"))).getTypeSignature());
-
-        assertEquals(
-                parseTypeSignature("map(my_enum(enum:varchar{\"k\": \"v)))\"}), my_enum_2(enum:bigint{\"k\": 1}))"),
-                new TypeSignature(
-                        StandardTypes.MAP,
-                        TypeSignatureParameter.of((new VarcharEnumType("my_enum", new VarcharEnumMap(ImmutableMap.of("k", "v)))"))).getTypeSignature())),
-                        TypeSignatureParameter.of(new LongEnumType("my_enum_2", new LongEnumMap(ImmutableMap.of("k", 1L))).getTypeSignature())));
-
-        assertSignatureFail("test_enum(enum:bigint{\"k\"})");         // no value
-        assertSignatureFail("test_enum(enum:bigint{\"k\", 2})");      // `,` instead of `:`
-        assertSignatureFail("test_enum(enum:bigint{})");              // empty map
-        assertSignatureFail("test_enum(enum:bigint{a: 2})");          // no quotes around key
-        assertSignatureFail("test_enum(enum:varchar{\"a\" \"b\"})");  // missing `:`
-        assertSignatureFail("test_enum(enum:varchar{:\"a\"})");       // missing key before `:`
-        assertSignatureFail("test_enum(enum:varchar{,\"a\"})");       // missing key before `,`
-        assertSignatureFail("test_enum(enum:bigint{{\"a\": 2})");     // extra `{`
-        assertSignatureFail("test_enum(enum:bigint{\"a\":: 2})");     // extra `:`
-        assertSignatureFail("t(enum:bigint{\"k\": {\"k1\": 1}})");    // nested enum
-        assertSignatureFail("test_enum(enum:bigint{\"k\": 2}haha)");  // extra input after `}`
-        assertSignatureFail("test_enum(enum:varchar{\"k\": 2})");     // long value for varchar enum
-        assertSignatureFail("test_enum(enum:bigint{\"k\": \"2\"})");  // varchar value for long enum
-        assertSignatureFail("test_enum(enum:bigint{\"k\": 2-})");     // invalid number
-        assertSignatureFail("test_enum(enum:bigint{\"k\": -})");      // invalid number
-        assertSignatureFail("test_enum(enum:bigint{\"k\": 2.29})");   // decimal value
-        assertSignatureFail("test_enum(enum:bigint{\"k\": \"2})");    // missing closing `"`
-        assertSignatureFail("test_enum(enum:varchar{\"k\": \"2\")");  // missing closing `}`
-        assertSignatureFail("test_enum(enum:bigint{\"k\": \"2\"}");   // missing closing `)`
-    }
-
     private static void assertRowSignature(
             String typeName,
             Set<String> literalParameters,
@@ -380,7 +330,7 @@ public class TestTypeSignature
     {
         try {
             parseTypeSignature(typeName);
-            fail("Type signature should fail to parse");
+            fail("Type signatures with zero parameters should fail to parse");
         }
         catch (RuntimeException e) {
             // Expected
@@ -391,7 +341,7 @@ public class TestTypeSignature
     {
         try {
             parseTypeSignature(typeName, literalCalculationParameters);
-            fail("Type signature should fail to parse");
+            fail("Type signatures with zero parameters should fail to parse");
         }
         catch (RuntimeException e) {
             // Expected
